# Devirtualize XmlEvents dispatch to give the optimizer full visibility