As `XmlEvents` is not public API any more, there should be no possibility of code bloat due to multiple instantiations and we can safely monomorphise dispatch from the tokenizer into the parser.

This yields small but measurable improvements in the benchmarks, but only after marking the single implementation of `XmlEvents::token` as `#[inline(always)]`, so possibly by avoiding one level of dispatching on the token type, c.f. before

```
test huge_roxmltree                         ... bench:   5,394,963 ns/iter (+/- 143,581)
test large_roxmltree                        ... bench:   2,646,532 ns/iter (+/- 19,663)
test medium_roxmltree                       ... bench:     698,086 ns/iter (+/- 70,686)
test tiny_roxmltree                         ... bench:       4,236 ns/iter (+/- 99)
```

and after

```
test huge_roxmltree                         ... bench:   5,167,046 ns/iter (+/- 227,797)
test large_roxmltree                        ... bench:   2,422,624 ns/iter (+/- 47,267)
test medium_roxmltree                       ... bench:     685,669 ns/iter (+/- 23,478)
test tiny_roxmltree                         ... bench:       3,936 ns/iter (+/- 56)
```