@adamreichold 2025-01-13
> I honestly would suggest passing `Context` directly. `XmlEvents` is just an artifact of old codebase.

I did that but the tokenizer tests still use the interface to collect raw streams of tokens. While it is an artifacts, this way of doing it allows us to reap the performance benefits without having to apply conditional compilation or some such for the tokenizer tests.