@RazrFalcon 2025-01-14
Passing `Context` directly would also make code a bit cleaner, but sure, we can do it later.