@adamreichold 2025-01-14
> Passing Context directly would also make code a bit cleaner, but sure, we can do it later.

It would allow to completely drop `Context::token` and theoretically even `enum Token`. But we then have to stub out a different `Context` definition and implementation for the tokenizer tests so they can still sort of collect tokens for their assertions.

I can start up an MR with those changes so we can have a look, but I don't think it should hold this simple performance win.